short to int

+++ initialize.h
@@ -521,7 +521,7 @@ struct sub_request{
  */
 struct read_hot{
 	unsigned int lpn;                  //这里表示该子请求的逻辑页号（该子请求的目标地址）
-	unsigned short num;                  //次数
+	unsigned int num;                  //次数
 	struct read_hot *next;
 };
 
@@ -531,9 +531,9 @@ struct read_hot{
  */
 struct seq_write{
 	unsigned int lpn;                  //这里表示该子请求的逻辑页号（该子请求的目标地址）
-	unsigned short read_num;          //读次数
-	unsigned short write_num;          //写次数
-	unsigned short hdd_num;          //写hdd次数
+	unsigned int read_num;          //读次数
+	unsigned int write_num;          //写次数
+	unsigned int hdd_num;          //写hdd次数
 	struct seq_write *next;
 };
